Extract folder file entry metadata
